Convert the Decimal positions and repin the final substrate
datom-codec 0.31.0 removed every datom kind from `f64`, so the four
`Decimal` positions in `horizon.ethos` now generate
`datom_codec::Decimal` rather than `f64`. The projection widens each
back to the `f64` the serialized view carries, which is lossless and
cannot produce a value the view could not hold, because a `Decimal` is
finite by construction. The contract test builds its literals through
the fallible constructor.

`datom-codec` stops being optional: the generated contract names
`datom_codec::Decimal` outside the `datom` gate, so the crate cannot
compile without it. The gate remains for the datom kinds themselves.
`datom-codec/rkyv` is enabled because `FixedLocation` is archived.

protos 0.31.0, datom-codec 0.31.0, ethos-zero 10.0.0.

+++ lib/src/projection/node.rs
@@ -2,6 +2,8 @@
 
 use std::collections::BTreeMap;
 
+use datom_codec::Decimating;
+
 use super::error::Error;
 use super::names::Named;
 use super::views::Projection;
@@ -267,11 +269,14 @@ impl NodeProjection for NodeDefinition {
             fixed_location: self
                 .fixed_location_option
                 .as_ref()
+                // A `Decimal` is finite by construction, so widening it back to
+                // the `f64` the serialized view carries cannot produce a value
+                // the view could not hold.
                 .map(|value| FixedLocationView {
-                    latitude: value.first_decimal,
-                    longitude: value.second_decimal,
-                    altitude: value.third_decimal,
-                    accuracy: value.fourth_decimal,
+                    latitude: value.first_decimal.float(),
+                    longitude: value.second_decimal.float(),
+                    altitude: value.third_decimal.float(),
+                    accuracy: value.fourth_decimal.float(),
                 }),
         })
     }
